Fixing the ARP table parsing
"Fixing" the warnings (unused variables) screwed up the ARP
table parsing, since the calls to strtok(NULL," ") where
taken out together with the variables, but the calls have
a side effect that is needed to correctly parse the ARP table.

Also the debug mode was changed to fork anyway and place the
debugging output into the syslog (with priority LOG_DEBUG) instead
of staying in the foreground and printing to stdout. This way the
boot process and all it's dependencies can run normally without
having to adjust it to the non-daemon character of parprouted when
debugging is switched on.